backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai>=1.0.0
youtube-transcript-api==0.6.2
pydantic==2.9.0
pydantic-settings==2.5.0
slowapi==0.1.9
python-dotenv==1.0.1
